refactor: update interfaces and logic where they used
