+++ src/app/core/models/api-response.model.ts
@@ -1,4 +1,4 @@
-export interface GetResponse<T> {
+export interface ApiResponse<T> {
   body: T;
   success: boolean;
   message: string;
