+++ src/app/core/models/heatmap-data.model.ts
