+++ src/app/core/models/poll-instance.model.ts
@@ -1,10 +1,9 @@
-import { UUID } from 'crypto';
 import { BaseModel } from './common/base.model';
 import { StudentModel } from './student.model';
 import { AnswerModel } from './answer.model';
 
 export interface PollInstanceModel extends BaseModel {
-  uuid: UUID;
+  uuid: string;
   finishedAt: Date;
   student: StudentModel;
   answers: AnswerModel[];
