+++ src/app/core/services/interfaces/cosmic-latte-poll-import-list.interface.ts
@@ -1,3 +1,5 @@
+import { BaseModel } from '../../models/common/base.model';
+
 export interface Audit {
   createdBy: string;
   modifiedBy: string;
@@ -38,13 +40,12 @@ interface Component {
   variables: Variable[];
   audit: Audit | null;
 }
-export interface PollInstance {
-  id: number;
+
+export interface PollInstance extends BaseModel {
   idCosmicLatte: string;
   uuid: string;
   name: string;
   version: string;
   finishedAt: string;
   components: Component[];
-  audit: Audit;
 }
