+++ src/app/core/services/interfaces/student.interface.ts
