+++ src/app/core/services/poll-instance.service.ts
@@ -3,7 +3,8 @@ import { Injectable } from '@angular/core';
 import { environment } from '../../../environments/environment';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { PollInstanceResponse } from './Types/pollInstance';
+import { ApiResponse } from '../models/api-response.model';
+import { PollInstanceModel } from '../models/poll-instance.model';
 
 @Injectable({
   providedIn: 'root',
@@ -23,8 +24,8 @@ export class PollInstanceService {
   getPollInstancesByFilters(
     cohortId: number,
     lastDays: number
-  ): Observable<PollInstanceResponse> {
-    return this.http.get<PollInstanceResponse>(
+  ): Observable<ApiResponse<PollInstanceModel[]>> {
+    return this.http.get<ApiResponse<PollInstanceModel[]>>(
       `${this.apiUrl}/${this.endpoint}/filter?cohortId=${cohortId}&days=${lastDays}`
     );
   }
